zephyr: include: sof: trace: Fix the trace level setting
Header has a guard to not reinclude logging header if it was already
included but inside this guard there is a definition of
SOF_ZEPHYR_TRACE_LEVEL which is later on used. So if log.h is already
earlier included SOF_ZEPHYR_TRACE_LEVEL is not defined and compilation
fails.

Moving this definition outside of the guard because if log.h is already
included LOG_LEVEL_INF is available.

Signed-off-by: Krzysztof Chruściński <krzysztof.chruscinski@nordicsemi.no>